requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
duckduckgo-search==7.3.2
requests==2.32.3
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0